Apply debian dosfstools-2.11-5 12-zero-slot.dpatch
----------
12-zero-slot.dpatch by Karl Tomlinson <karlt@karlt.net>

Fixes crashes due to zero slot numbers causing a negative offset in the
call to copy_lfn_part in lfn_add_slot. On amd64 this results in a SIGSEGV
in copy_lfn_part. On x86 the result is heap corruption and thus sometimes
a SIGSEGV or double free abort later.
(Closes: #152550, #353198, #356377, #401798).
----------

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>