dosfsck: fix truncate size of circular cluster chain
In case that fsck found circular cluster chain, it tries to truncate that
file.

When the -t option is given, fsck keep well cluster size before circular
cluster chain is founded. But when the -t option is not given, fsck does not
keep cluster size.

Modify it to keep the cluster size, even if the -t option is not given.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>