dosfsck: fix truncate size in case of shared clusters
In case that two files share a cluster, fsck tries to truncate one of those
files. In general(auto mode), fsck choose second files to truncate size
just before the shared cluster.

Now fsck choose second file, but truncate it with size of first file.
Modify it to truncate second file with size of that.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>