dosfsck: exclude boot dirty field from comparison
Dirty flag field in backup boot sector is not handled in linux.
But dosfsck always compare with backup boot sector.

So, once dirty flag are different in boot and backup boot, it can't be
fixed.
(because boot sector's difference are not corrected automatically.)

This patch exclude boot dirty field from comparison with boot and backup
boot.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>