dosfsck: make FOUND.XXX directory for reclaimed files
Apply FOUND.XXX
All reclaimed files were created in root directory.
Their name format is FSCKXXXX.REC, XXXX is integer value 0 to 9999.
So, if there are more than 9999 reclaimed files in one directory,
fsck can't create reclaimed file and will fail to execute.

Now, they are located in FOUND.000 directory with name of FSCKXXXX.REC.
So, if there are more than 9999 files in one directory, it make
FOUND.001 and create files in that directory.

Auto correcting
If filesystem information sector(fsinfo sector) is 0, it wll make initial
fsinfo and write to 1-th sector and set to fsinfo field in boot sector.

If there is an entry which can't recognize file or directory, fsck
consider it to file.

If boot sector is corrupted, check backup boot sector and if backup is
valid, copy it to original.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>