dosfsck: fix free first FAT entry of root directory
If first FAT entry of root direcotry had been set to zero,
dosfsck of previous version will be terminated.

So, in this case, that is, first FAT entry of root directory had been set to
free, just set EOC(End of cluster) to first FAT entry of root, not terminated.
Terminate only when first FAT entry of root directory is BAD.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>