io: reimplement fsync() as sync_file_range()
If modification due to fsck(dosfsck) are heavy, fsync() takes so much time
at the end of program. It means fsck process is in uninterruptible sleep
status and process can't receive signal to terminate or something.

So, reimplement fsync() as sync_file_range(), which divide IO requests
and waiting for completion into 512M units.

Change-Id: Icd4d9b042d0a8d7581a1db3e01575d17001939a4
Signed-off-by: JaeHoon Sim <jay.sim@lge.com>